insert text with autofocus

# function is identical to textAreaInput but adds autofocus to tags

text_focus(inputId, label, value = "", width = NULL, height = NULL,
  cols = NULL, rows = NULL, placeholder = NULL, resize = NULL)

Arguments

inputId

The input slot that will be used to access the value.

label

Display label for the control, or NULL for no label.

value

Initial value.

width

The width of the input, e.g. '400px', or '100%'; see validateCssUnit.

height

The height of the input, e.g. '400px', or '100%'; see validateCssUnit.

cols

Value of the visible character columns of the input, e.g. 80. If used with width, width will take precedence in the browser's rendering.

rows

The value of the visible character rows of the input, e.g. 6. If used with height, height will take precedence in the browser's rendering.

placeholder

A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.

resize

Which directions the textarea box can be resized. Can be one of "both", "none", "vertical", and "horizontal". The default, NULL, will use the client browser's default setting for resizing textareas.